test: Move GET /observations to its own test file
